# Rework MapStr.walkMap to support dotted keys