Currently if a dotted key is passed to walkmap it is split up in it's part and the map is traversed for each part. In case an event has a dotted key inside walkMap does not return the expected value. For the Event `{"a.b": 1}` the method `GetValue("a.b")` has previously returned `ErrKeyNotExists` but now returns 0.

With this change all processors which rely on Get/Put functionalities of MapStr can now also work with events containing dotted keys.

**Changes**

* The walkMap function was rewritten to a recursive function. This made adding the new functionality easier.
* Add tests for GetValue to make possible usages more clear
* Add Benchmark tests for walkMap use cases

**Limitations**

When using `walkMap` with a Key, either the full key has to be dotted or the last part. `GetValue("a.b.c")` on an event `{"a.b":{"c":17}}` does not work as only the first part of the key or the full key are checked against the map.

Benchmark old to new walkMap:

```
go test  -bench=BenchmarkWalkMap -run=BenchmarkWalkMap
```

Recursive:
```
BenchmarkWalkMap/Get-4           1000000              1274 ns/op
BenchmarkWalkMap/Put-4           1000000              1277 ns/op
BenchmarkWalkMap/HasKey-4        2000000               880 ns/op
BenchmarkWalkMap/HasKeyFirst-4  20000000               111 ns/op
BenchmarkWalkMap/Delete-4        1000000              1313 ns/op
```

Iterative:
```
BenchmarkWalkMap/Get-4            500000              2608 ns/op
BenchmarkWalkMap/Put-4           1000000              1140 ns/op
BenchmarkWalkMap/HasKey-4        3000000               411 ns/op
BenchmarkWalkMap/HasKeyFirst-4  20000000                85.9 ns/op
BenchmarkWalkMap/Delete-4        1000000              1008 ns/op
```

The benchmark show a difference in the code before and after. The `GetValue` function has become quite a bit more efficient. The `HasKey` is now less efficient. I still need to investigate in detail where the differences come from but the assumption is it comes from the additional functionality.

The old function was left in the code for comparison reason in this PR. It can be removed later.

This change was initially triggered by https://github.com/elastic/beats/pull/5916 but is split out into a separate PR because it has usage in multiple places.