@andrewkroh 2018-02-19
Could you run the benchmark with -benchmem enable.