@ruflin 2018-02-19
Here is the output:

Recursive
```
BenchmarkWalkMap/Get-4           1000000              1263 ns/op             432 B/op          7 allocs/op
BenchmarkWalkMap/Put-4           1000000              1329 ns/op            1120 B/op         10 allocs/op
BenchmarkWalkMap/HasKey-4        2000000               846 ns/op             192 B/op          6 allocs/op
BenchmarkWalkMap/HasKeyFirst-4  10000000               110 ns/op              32 B/op          1 allocs/op
BenchmarkWalkMap/Delete-4        1000000              1324 ns/op            1120 B/op         10 allocs/op
```

Iterative
```
BenchmarkWalkMap/Get-4            500000              2677 ns/op             824 B/op         13 allocs/op
BenchmarkWalkMap/Put-4           1000000              1015 ns/op            1072 B/op          8 allocs/op
BenchmarkWalkMap/HasKey-4        5000000               375 ns/op              96 B/op          2 allocs/op
BenchmarkWalkMap/HasKeyFirst-4  20000000                83.5 ns/op            16 B/op          1 allocs/op
BenchmarkWalkMap/Delete-4        1000000              1038 ns/op            1072 B/op          8 allocs/op
```