@ruflin 2018-02-19
Comparison with the latest changes

Recursive
```
BenchmarkWalkMap/Get-4           1000000              1334 ns/op             432 B/op          7 allocs/op
BenchmarkWalkMap/Put-4           1000000              1273 ns/op            1120 B/op         10 allocs/op
BenchmarkWalkMap/HasKey-4        2000000               690 ns/op             160 B/op          5 allocs/op
BenchmarkWalkMap/HasKeyFirst-4  50000000                28.8 ns/op             0 B/op          0 allocs/op
BenchmarkWalkMap/Delete-4        1000000              1431 ns/op            1120 B/op         10 allocs/op

```

Iterative
```
BenchmarkWalkMap/Get-4            500000              2577 ns/op             824 B/op         13 allocs/op
BenchmarkWalkMap/Put-4           1000000              1052 ns/op            1072 B/op          8 allocs/op
BenchmarkWalkMap/HasKey-4        5000000               381 ns/op              96 B/op          2 allocs/op
BenchmarkWalkMap/HasKeyFirst-4  20000000                89.6 ns/op            16 B/op          1 allocs/op
BenchmarkWalkMap/Delete-4        1000000              1083 ns/op            1072 B/op          8 allocs/op
```