@urso 2018-02-20
@ruflin for comparing benchmarks, have a look at [benchcmp](https://godoc.org/golang.org/x/tools/cmd/benchcmp). 

I believe most of the difference in performance comes from the `splitN(, 2)` vs `split` call. I think this can be optimised even further by using an iterative approach (tail-call optimization not done by compiler) and using `strings.Index` to work on sub-strings of the key, so to reduce allocations even further.