@ruflin 2018-02-23
@urso Thanks for the tool. Will use it next time.

For the optimizations: Put it on my todo list as I think what you suggest is definitively worth it as we call this function a lot.